WW-3714 Deprecate and migrate AbstractInterceptor and MethodFilterInterceptor
