+++ core/src/main/java/org/apache/struts2/interceptor/AbstractInterceptor.java
@@ -16,28 +16,46 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.opensymphony.xwork2.interceptor;
+package org.apache.struts2.interceptor;
 
-import com.opensymphony.xwork2.ActionInvocation;
+import org.apache.struts2.ActionInvocation;
 
 /**
- * @deprecated since 6.7.0, use {@link org.apache.struts2.interceptor.AbstractInterceptor} instead.
+ * Provides default implementations of optional lifecycle methods
  */
-@Deprecated
-public abstract class AbstractInterceptor extends org.apache.struts2.interceptor.AbstractInterceptor implements ConditionalInterceptor {
+public abstract class AbstractInterceptor implements ConditionalInterceptor {
+
+    private boolean disabled;
+
+    /**
+     * Does nothing
+     */
+    public void init() {
+    }
+
+    /**
+     * Does nothing
+     */
+    public void destroy() {
+    }
 
     /**
      * Override to handle interception
      */
     public abstract String intercept(ActionInvocation invocation) throws Exception;
 
-    @Override
-    public String intercept(org.apache.struts2.ActionInvocation invocation) throws Exception {
-        return intercept(ActionInvocation.adapt(invocation));
+    /**
+     * Allows to skip executing a given interceptor, just define {@code <param name="disabled">true</param>}
+     * or use other way to override interceptor's parameters, see
+     * <a href="https://struts.apache.org/core-developers/interceptors#interceptor-parameter-overriding">docs</a>.
+     * @param disable if set to true, execution of a given interceptor will be skipped.
+     */
+    public void setDisabled(String disable) {
+        this.disabled = Boolean.parseBoolean(disable);
     }
 
     @Override
     public boolean shouldIntercept(ActionInvocation invocation) {
-        return shouldIntercept((org.apache.struts2.ActionInvocation) invocation);
+        return !this.disabled;
     }
 }
