+++ core/src/test/java/com/opensymphony/xwork2/config/providers/ConfigurationProviderOgnlAllowlistTest.java
@@ -61,6 +61,7 @@ public class ConfigurationProviderOgnlAllowlistTest extends XWorkJUnit4TestCase 
                 Class.forName("com.opensymphony.xwork2.mock.MockInterceptor"),
                 Class.forName("com.opensymphony.xwork2.Action"),
                 Class.forName("com.opensymphony.xwork2.interceptor.AbstractInterceptor"),
+                Class.forName("org.apache.struts2.interceptor.AbstractInterceptor"),
                 Class.forName("com.opensymphony.xwork2.Result"),
                 Class.forName("com.opensymphony.xwork2.SimpleAction"),
                 Class.forName("org.apache.struts2.interceptor.Interceptor"),
@@ -92,6 +93,7 @@ public class ConfigurationProviderOgnlAllowlistTest extends XWorkJUnit4TestCase 
                 Class.forName("com.opensymphony.xwork2.mock.MockInterceptor"),
                 Class.forName("com.opensymphony.xwork2.Action"),
                 Class.forName("com.opensymphony.xwork2.interceptor.AbstractInterceptor"),
+                Class.forName("org.apache.struts2.interceptor.AbstractInterceptor"),
                 Class.forName("com.opensymphony.xwork2.Result"),
                 Class.forName("com.opensymphony.xwork2.SimpleAction"),
                 Class.forName("org.apache.struts2.interceptor.Interceptor"),
@@ -123,6 +125,7 @@ public class ConfigurationProviderOgnlAllowlistTest extends XWorkJUnit4TestCase 
                 Class.forName("com.opensymphony.xwork2.Validateable"),
                 Class.forName("com.opensymphony.xwork2.Action"),
                 Class.forName("com.opensymphony.xwork2.interceptor.AbstractInterceptor"),
+                Class.forName("org.apache.struts2.interceptor.AbstractInterceptor"),
                 Class.forName("com.opensymphony.xwork2.Result"),
                 Class.forName("org.apache.struts2.interceptor.Interceptor"),
                 Class.forName("org.apache.struts2.interceptor.ConditionalInterceptor"),
